コルーチンを利用して OkHttpClient を非 main スレッドで初期化する
